+++ labs/Lab_2_chatWithDocuments/deploy/azure-deploy.json
@@ -61,7 +61,7 @@
     },
     "packageUri": {
       "type": "string",
-      "defaultValue": "https://github.com/Microsoft-USEduAzure/OpenAIWorkshop/raw/main/labs/Lab_2_chatWithDocuments/orchestrator/orchestrator-func-app/orchestrator.zip",
+      "defaultValue": "https://github.com/Microsoft-USEduAzure/OpenAIWorkshop/raw/main/labs/Lab_2_chatWithDocuments/orchestrator/orchestrator.zip",
       "metadata": {
         "description": "The zip content url."
       }
