+++ labs/Lab_2_chatWithDocuments/orchestrator/.funcignore
@@ -1,5 +1,9 @@
 .git*
 .vscode
-local.settings.json
+__azurite_db*__.json
+__blobstorage__
+__queuestorage__
 test
 .venv
+.env
+local.settings.json
