+++ labs/Lab_2_chatWithDocuments/orchestrator/.gitignore
@@ -133,4 +133,3 @@ __blobstorage__
 __queuestorage__
 __azurite_db*__.json
 .python_packages
-
