+++ labs/Lab_2_chatWithDocuments/orchestrator/host.json
@@ -10,10 +10,6 @@
   },
   "extensionBundle": {
     "id": "Microsoft.Azure.Functions.ExtensionBundle",
-    "version": "[3.*, 4.0.0)"
-  },
-  "concurrency": {
-    "dynamicConcurrencyEnabled": true,
-    "snapshotPersistenceEnabled": true
+    "version": "[4.*, 5.0.0)"
   }
 }
